stm:fdma Allow transfer descriptors to be re-used
The dmaengine will automatically retire a completed transfer to the free
list ready for re-use when the flag DMA_CTRL_ACK is specified in the
transfer flags. If this flag is not specified then the the transfer is
moved to the free list, but cannot be re-used until DMA_CTRL_ACK is set.

This basically implies that assuming the transfer flags don't request any
source or destination memory unmapping etc, that you can re-use the same
transfer to do the same memory DMA operation again.

This patch makes the neccessary changes to the FDMA internal descriptor
handling to allow this. Primarily this is ensuring that the internal lists
do not become corrupt and that the transfer is correctly placed in the free
list once it is finally ACKed.

Signed-off-by: John Boddie <john.boddie@st.com>