stm: pio-control: Remove support for don't care values in retiming fields
The retiming code has always supported a field value of -1, which
indicated that the current value in the hardware should not be
changed.  It should have only been used to indicate that we don't care
what value a register should take (to match the "X" notation used in
some of the application notes), but its not clear that this rule is
always being followed.

It also means that the result of a particular configuration may depend
on the previous state, and so has the potential to cause hard to
diagnose problems.

So remove this support entirely, and where existing macros set the
fields to -1 replace this with 0.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>