stm: pio-control: Move retiming paramters into pio config
Currently the retiming parameters need to be passed into all functions
which need them. Now that the retiming style is part of the PIO control
configuration, it makes sense to move the retiming parameters into the
data structure as well. This reduces the number of parameters which
have to be passed in, makes it easier to add extra parameters, and
would make it easier to support devices with multiple retiming
styles.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>