stm: pio-control: Pass pad direction into retiming functions
Prior to this change the retiming code has not been aware of which
direction a pad was operating in. However future changes would make
this information useful:

 - in the "packed" retiming style the delay is different on the input and
   ouput paths, which means if we want to handle the dalay as a time
   period rather than a bit pattern the code needs to know the direction.

 - in the "dedicated" retiming style the delay is only supported on the
   input or output path, so knowing which direction the pad is configured
   would save the user having to specify this information.

When configuring a pad, we know the direction as it passed in by the
caller. However when generating debugfs output the code has to read the
direction from the hardware.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>