stm: arm: make DEBUG_LL UART choice board dependant
The UART to be used as console (including DEBUG_LL support) should depend
on the board instead of the SoC type, as we could have more boards
equipped with the same SoC where the UARTs are different.

Clearly the constraints forced by DEBUG_LL make impossible to have
multiple boards support working in the same kernel, but its usage is usually
done in the bring-up phase or to debug issues in the early debug stage,
where multiple boards support is not a problem. In any case it's worth to
improve it a bit.

This patch fixes the UART for the b2000 (Orly-1) board also.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>