stm: mpe41: Fix compile error in PMU setup
This patch fixes a compilation error introduced in this commit:

ed3d79d6 "stm: Move STxMPE41 PMU support to ARM specific directory"

Unfortunately the Newman and Orly do not use the same macros to
access sysconfigs, so we define a SYSCONFIG macro for the Newman
to allow the code to work. There really needs to be a common
macro to access syconfigs on the MPE41.

Signed-off-by: Chris Smith <chris.smith@st.com>
Signed-off-by: David McKay <david.mckay@st.com>