stm: asc: Don't enable default console when another has been specified
Most users specify console= in their kernel parameters, but if it isn't
given then we typically put the console on an ASC at the default baud,
9600. This is fine, but there is currently a bug whereby if only a
non-ASC console is specified in the kernel parameters then the default
console is enabled /as well/ and writing to a serial port at 9600 baud
cripples our performance any time there is a significant amount of
console output.

This patch prevents the default ASC console being enabled if console=
was given. This means that if only a non-ASC console is specified, it
becomes the sole console. Of course, the option remains to specify
several consoles. Users specifying either an ASC console, or no console,
see no change.

Signed-off-by: Chris Smith <chris.smith@st.com>