stm: Improve secondary boot error handling
You can't boot the secondary processor if the primary processor failed
to create the identity mapping. This patch introduces a check for this
and prints a warning so the user knows what has gone wrong.

It also zeros out the secondary_data.swapper_pg_dir field in the event
of failure for completeness, as this field was added later.

Signed-off-by: David McKay <david.mckay@st.com>