stm:snd Enable SPDIF_FORMATING after enabling the uniperif player
Even if you enable SPDIF_FORMATING and the player with an atomic 32bit
write it can still hit the silicon bug which hangs the player. i

The silicon bug is that enabling the SPDIF formatting control bit before
turning on the player control bit can result in the hardware hanging.
Writing both bits in the same 32 bit write can also cause this lockup.
This code ensure that the player control bit is always turned on before
the spdif formatting bit.

The spinlock ensures we aren't preempted between enabling the player and
enabling spdif formating.

Signed-off-by: Peter Griffin <peter.griffin@st.com>