stm:miphy: Cleanup the callbacks
This patch clean's up the callbacks in miphy structure, so that device
tree stuff can fit in neatly with others.

Patch changes mp_select to take data pointer which could be a device
state or device config itself, and the function can access the sysconfs
via that pointer.
It adds new Init and exit calls to the mp platform data, so that and
Init can claim the resources and set the sysconfs accordingly and the
exit can release them gracefully.

Going this way will help both device tree integration and also clean up
the existing code.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>