stm:snd Fix compilation error when CONFIG_PM_RUNTIME not defined
This patch fixes a "'struct  dev_pm_info' has no member named
'runtime_status'" compilation error in the uniperipheral drivers by
removing the checks on whether the device is already active or
suspended.

The SNDRV_PCM_INFO_RESUME flag has also been removed as this is for
resuming the IP where it was initially paused, and this is not yet
supported.

Also, the 'thaw' function is now a 'restore' function ensuring the
correct PM callbacks are used.

To prevent lock-ups calling the FDMA during suspend, we no longer check
if parking mode is active, instead relying on the check for the
uniperipheral still running.

Signed-off-by: John Boddie <john.boddie@st.com>