stm:snd:stih415 Add power management support to ALSA glue driver
This patch adds power management support to the STiH415 ALSA glue driver.
To do this the code has changed from just a loadable module to an actual
platform driver allowing the power management callbacks to be added.

The power management callbacks are required as the sysconf registers do
not retain there value after HoM.

Signed-off-by: John Boddie <john.boddie@st.com>