stm_nand_afm: improve the empty-page test
The AFM and boot-mode ECC schemes are such that the ECC data generated by
reading an empty (erased) page would normally be interpreted as the result of an
"uncorrectable" error.  To avoid this issue, an additional check is performed to
test for the empty-page condition.

Previously, the test involved checking that the ECC data (both the stored and
calculated values) were consistent with an empty page.  However, certain
non-empty patterns of data produce the same ECC results and could be incorrectly
identified as empty pages, rather than throwing uncorrectable errors.  In most
cases, this problem does not arise; uncorrectable errors should not occur if
operating within the specs of the NAND device.  However, it helps to detect data
that may have been generated by an "alien" driver if we return the uncorrectable
error condition in this situation.

This patch strengthens the empty-page test by also examining the page data.  For
most read operations, this adds little or no overhead, since the check will
break early on.  For reading empty pages, we must now check the all page data,
but typically this would be done only once, for example, when mounting a
filesystem.

Signed-off-by: Angus Clark <angus.clark@st.com>