stm_nand: add STM_NAND_SAFE_MOUNT option to detect 'alien' BBTs
When scanning for factory-programmed bad block markers or Flash-resident bad
block tables (BBTs), it is assumed that the NAND device has either not been used
previously, or used only with the same NAND driver and set of driver options.

If switching from one NAND driver to another (or one set of driver options to
another), it is important to ensure the device is 'cleaned' properly first.
Failure to do so can cause a number of problems, including:
    - Uncorrectable ECC errors when reading BBTs
    - Programmed good blocks falsely interpreted as bad
    - Corrupted BBTs written to Flash
    - Manufacturer-programmed bad block markers permanently lost
    - Information about worn blocks permanently lost

To help avoid these issues, this patch introduces the 'STM_NAND_SAFE_MOUNT'
option.  During driver initialisation, an attempt is made to detect 'alien' BBTs
that may have been left behind by a previous driver.  If any alien BBTs are
detected, the driver will abort scanning for bad block markers and will make no
attempt to read or write BBTs.  The device is then mounted in a "Recovery" mode,
where all blocks are treated as bad, preventing any corruption or loss of data,
but still allowing some limited access to assist with the analysis and recovery
of the device.

The patch is intended to work in conjunction with the 'STM NAND Tools' package
which can help with cleaning a device, synchronising bad block markers with the
BBT data, and migrating BBTs from one format to another.

Alternatively, the kernel can be re-built with the reported 'alien' driver, and
then used to clean the device properly before switching back to the intended
driver.

Given the severity of the potential problems, the STM_NAND_SAFE_MOUNT option is
enabled by default.  However, the option adds some overheads to driver
initialisation and so should be disabled on production systems, once the driver
selection has been finalised and there is no danger of encountering 'alien'
data.

Signed-off-by: Angus Clark <angus.clark@st.com>