mtd: mtdchar: ECCGETLAYOUT: report the real number of eccbytes
Commit cc26c3cd3d1cf40a07f2b19ac4c53d517bee52a5 "mtd: nand: expand
nand_ecc_layout, deprecate ioctl ECCGETLAYOUT" introduced an expanded
'nand_ecc_layout' structure, to accommodate newer devices with larger OOB sizes.

To maintain backwards compatibility with the ECCGETLAYOUT IOCTL, the previous
nand_ecc_layout was retained in the user-space ABI as 'nand_ecc_layout_user',
and the IOCTL was updated to convert from the new structure to the old
structure, possibly truncating the layout information along the way.

Since the the ECCGETLAYOUT IOCTL was already marked for deprecation, no attempt
was made to add a new IOCTL to support the expanded nand_ecc_layout structure.

However, the STM NAND Tools package uses the ECCGETLAYOUT IOCTL, and in
particular the 'eccbytes' field, to determine the ECC mode when using the
stm-nand-bch driver.  Unfortunately, as part of the conversion process, eccbytes
gets capped to 'MTD_MAX_ECCPOS_ENTRIES', rather than reporting the real eccbytes
figure.

This patch restores the eccbytes field to reflect the real number of bytes used.
This should be regarded as a temporary measure.  A better means of determining
the BCH ECC mode will be investigated for the future.

Signed-off-by: Angus Clark <angus.clark@st.com>