stm_spi_fsm: add support for 32-bit addressing
This patch adds support for 32-bit addressing to the STM SPI Serial Flash
driver.  Although conceptually quite simple, there are a couple of issues that
complicate matters (hence, the rather verbose commit log!).

Firstly, it seems that different manufactures have adopted slightly different
approaches for supporting 32-bit addressing.  For example, the 'n25q256' device
employs an additional set of Read commands for 32-bit addresses, while Write and
Erase operations require the device to first enter a special 32-bit mode.  The
'mx25l25635e' device, on the other hand, requires entering a 32-bit mode prior
to all Read, Write, and Erase commands.  To complicate matters further, the
command sequence used to enter 32-bit mode also differs between the two devices.

The other issue that needs to be considered is what happens following a warm
reset (e.g. a watchdog reset) on a boot-from-spi system.  The SPIBoot controller
assumes that the Flash device is in its default reset state on boot, including
24-bit address mode, and is ready to accept commands.  This can be achieved by
using some form of on-board logic or controller to force a device POR in
response to a SoC-level reset, or by making use of the device reset signal if
available (limited to a handful of Macronix devices at present).

Failure to take such precautions can cause problems following a warm reset.  For
some operations, like Erase, there is little that can be done.  However, in the
case of 32-addressing, different options are often available, and these can help
minimize the window in which a reset could cause a problem.  For example, on the
'n25q256' device, we use the dedicated 32-bit address Read commands which do not
require leaving 24-bit mode.  For the other commands, and for the 'm25l25635e'
device, we enter 32-bit address mode only for the duration of the operation, and
then return to 24-bit address mode.

Alternatively, if the boot-device is not SPI, or if the Flash device is properly
reset following a SoC-level reset, then we simply enable 32-bit mode pervasively
during initialisation.  On non-boot-from-spi systems, the driver will take care
of the addressing mode when the device is probed.  On boot-from-SPI systems, we
make use of the fact that a reset will cause the device to return to 24-bit
address mode, ready for the SPIBoot Controller.

To this end, we introduce three new platform properties "reset_signal",
"reset_por", and "boot_from_spi".  The driver is updated to add a "Enter/Exit
32-bit address mode" sequence (fsm_seq_en32bitaddr), which gets configured
during initialisation according to the device and platform properties.  The
Read, Write, and Erase sequences are configured to use either 24-bit or 32-bit
addressing as appropriate.  In both cases, for simplicity, the address is now
split across the two available address registers, and issued using the ADDR1 and
ADDR2 instructions.  The only difference being the number of address bits issued
for ADDR1: 8 bits for 24-bit mode and 16 bits for 32-bit mode; ADDR2 issues the
remaining 16-bits in both cases.

Note, newer versions of the SPI-FSM IP support 32-bit addresses natively; a
single ADDR register/instruction can generate the 32-bit address.  However, this
would complicate the code somewhat and offers no real benefit for our driver.
One potential exception would be the "repeat sequence" feature, which involves
auto-address-incrementing, but reliability concerns have prevented its use so
far.

Signed-off-by: Angus Clark <angus.clark@st.com>