stm_spi_fsm: workaround for FIFO direction delay
The SPI FSM controller uses a single FIFO for both read and write operations,
with a configuration bit to set the direction as required.  In the case of write
operations, we need to set the FIFO to write mode, write the data to the FIFO,
then load/run the FSM program to output data to the Flash device.

However, it turns out that the IP requires a number of clock cycles for a
direction change to be completed.  Previously, a register readback seemed to
provide a sufficient delay.  One theory was that the register read would block
until the direction change had been completed.  However, problems have now been
reported when working at slower clock rates.

The issue is due to be investigated further by the Design Team, but the current
advice is to use a delay equivalent to 10 clock cycles between changing the FIFO
direction and writing to the FIFO.

This patch borrows a code segment from u-boot, based partly on this advice, and
partly on empirical tests.  The linux version has been tested up to a clock
divider of 50 (i.e. down to 2MHz SPI clock freq) which should be sufficient for
all sensible devices.

Signed-off-by: Angus Clark <angus.clark@st.com>