stm_mmc: No retiming delay on MMC for Orly/Lille
It doesn't make much sense to use retiming delays for a bidirectional
signal. Indeed later versions of the retiming hardware actually do not
permit this. The reworked retiming code now produces a warning if you
attempt to set a delay on a signal without knowing the direction, so
lots of warnings are produced on boot without this patch.

This patch removes the retiming on the MMC pads for the stih415 and the
stxh205.

Signed-off-by: David McKay <david.mckay@st.com>