stm-of: implement generic callbacks for stm devices.
This patch add callbacks for drivers which require some kind of
callbacks when using device trees.
Most of these callbacks are generic in some sense, and can be extended
if required. However, if some of the SOC's callbacks do not fit this
flow, a new soc level callbacks can be implemented.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>