stm: b2078: add support for SASC-1 Docsis3 reference board
This patch adds support for the SASC-1 "Docsis3" Reference Board b2078
it includes also the FastPath driver, but it does not seem to work on
this board.

Following IP/drivers has been tested: ASC (UART1), Leds, USB (with the
ethernet dongle), I2C and the SPI-NOR device.

*Note*
In order to connect, it was necessary to reduce the JTAG speed
(from 50000000 to 3000000, using the b2078 tp v05).

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>