stm: fdma: removed PAUSING state in the channel fsm
Looking at the channel finite state machine there is a
PAUSING state which is unreachable.

From PAUSING the fsm can move on several states
but there is no state which can move to PAUSING.
Threfore it's a dead state and it can be removed with
no impact in the channel fsm.

Tested on b2000-h416 with "DMA Test client" module

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>