stm_nand_afm: relax the empty-page test
Commit ee5f2413aa0b7652236710ac7bdf8efa3debe010 "stm_nand_afm: improve the
empty-page test" strengthened the empty-page test by requiring that all page and
OOB data be in the unprogrammed state, 0xff.  However, this turns out to be a
little too strict for the case of JFFS2 where a cleanmarker node is placed in
the OOB area to denote the block has been erased and is ready for use.  In
practice, we only need to test the ECC data in the OOB area.  Since this check
is already performed, we can simply remove the OOB data from the "at 0xff" test.

Signed-off-by: Angus Clark <angus.clark@st.com>