stm: stih415: remove useless tx_coe setting from platform
This patch removes a useless setting for the GMAC
in Orly1. In this SoC the MAC doesn't support the csum
insertion for the frame to be transmitted. This is
solved at run-time by looking at the HW feature
register. For the platforms w/o this HW feature
register the tx_coe must be passed from the
platform.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>