stmmac: manage tx clean process by NAPI
The process to clear the tx resources returns to be
managed by poll method as in the past. So this patch
removes the tasklet previously introduced. Calling
the tx clean routine inside the NAPI poll method is
better helping on mitigating the number of interrupts.

This patch also aligns the driver to the version
pushed in net-next git.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>