stm: mali: fix section mismatch warnings
Mali driver was producing several section mismatch warnings
as some no __init/__exit functions were referring respectively the
__init mali_platform_init and __exit mali_platform_deinit.

One option was adding __init/__exit annotation to all functions in the
chain (they were only called at init and exit time), but this did not
solve the issue completely as the mali_platform_deinit is also called
in an error path of mali_platform_init, so it means that it has valid usage outside the exit section.

So the best solution would be to remove the annotation from
mali_platform_init and mali_platform_deinit.

(IMO it's a symptom of a not clean design of the MALI interfaces)

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
CC: Angelo Castello <angelo.castello@st.com>