isve: manage and skip docsis header from the incoming frames
For some downstream queues it could be needed
to skip some bytes (DOCSIS Header) from the incoming frames.

In this case from the platform we can pass this information
to the driver that will be able to skip this header and
pass to the upper-layer the real Eth frame.

In case of the DOCSIS HW is able to pass frames to Linux
w/o any extra bytes so from the platform it won't be
necessary to fill this field at all.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>