mtt: add a sample application and driver
This adds sample code to the 'sample/mtt' folder.
The application does a blocking read on a sample module instrumented with
the MTT API. the purpose is to demonstrate how generated data can help
indentify a simulated hadware contention issue with help of the TimeCharting
tools (like Kptrace in STWorkbench).

Signed-off-by: Marc Titinger <marc.titinger@st.com>