Ensure pmb_calc does works for multiple mappings
Patch developed by Stuart Menefy and posted on Bugzilla 21658 comment 5
and also found necessary for Bugzilla 24459.  In both cases these
are related to use of the config option: CONFIG_PMB_LARGE_UNCACHED_MAPPING
which appears to be necessary to provide  this issue.  While this
patch is STAPI speciifc, the underlying issue is that the routine
takes a int * and, if passed the address of a single int when
the mapping requires 2 entries in the pmb - the pointer will be incremented and
garbage read on the second allocation.

Signed-off-by: Jon Frosdick <jon.frosdick@st.com>