isve: fix unnecessary calls to netif_carrier_off
we cannot call the netif_carrier_off when close
the interface unless to turn it on when re-open
it again otherwise no frames will be passed from
the layer.
So this patch removes this call to the netif_carrier_off
that for a virtual interface can be completely useless.

See bugilla bug 25031

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>