isve: move internal res allocation inside the probe
This patch allocates the resources inside the probe and releases them
when remove the driver.

Indeed it provides a work-around to a strange phenomenon when
put down/up the interface via ifconfig at runtime.
In fact, although the driver always releases its resources and
reallocates them again during the new open, it seems that the
DFWD_used_add register is not properly updated by the HW and in the
reception process the DFWD register 0x110 shows corrupted addresses.
This maybe should be debugged with docsis support team.
In any case, this patch allows us to put down/up the iface and
that means that the problem still remains when remove isve as module
and reinsert it with insmod.

This patch also:

o moves calls to ioremap and request_mem_region to
  devm_request_and_ioremap call;
o removes broked free_irq call;
o fixes the Kconfig DEBUG option when the driver is not built-in.

See bugilla bug 25031

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>