stm: Common up FDMA and coprocessor ELF loading code
We common up the ELF segment loading code used by the coprocessor and FDMA
drivers.  The intention is to make the loading code common to all physical
ELF loading schemes so that it can be enhanced with ST ELF loading
extensions such as Zlib compressed segments.

We support existing ioremap ranges via the ELF<xx>_LoadParams.  These will
be used in preference to creating new mappings if they cover the ranges
needed.  Drivers should not tell the loader about existing regions unless
they wish them to be used for loading.

ELF files must have the correct SoC physical addresses in the ELF segment
(program header) PADDR (p_paddr) fields.

Note: I have not tried to address the problem that if the coprocessor BPA2
partition is within kernel-known RAM when built for ARM, the coprocessor
ELF will fail to load as the RAM cannot be ioremapped.  Before this patch
the load would fail silently.  With this patch it no longer fails silently.

Signed-off-by: David Cook <david.cook@st.com>