stm: libelf: Allow load using cached/write-combined mappings if the arch allows
Only the ST40 allows cached mappings so far; the ARM and mnay other architectures
have no cache flush function to go with ioremap_cached().

For other architectures, when cached mappings are requested we use ioremap_wc().
When cached is not requested, write-combining will not be used either (just in
case it's possible that some memories will not work correctly with
write-combining).

Signed-off-by: David Cook <david.cook@st.com>