stm: libelf: Add support for STM extended ELF with Zlib compressed segments
We have a Zlib compressed flag on ELF sections and segments if a segment
is compressed.  The p_memsz contains the uncompressed size of the segment
and the p_filesz the compressed size.

If before compression the p_filesz was less than p_memsz the compressed
data will include the zero-padding required to complete the load.

The zlib decompression function zlib_inflate_blob() was hard-coded to not
expect a zlib data header, which our extended ELF format does have, so we
add another zlib decompression function zlib_inflate_blob_with_header().

Signed-off-by: David Cook <david.cook@st.com>