stm: libelf: Add support for STM extended ELF with checksums
We add support for ST ELF auxiliary segments/sections which can contain
information on CRC32 and Adler32 checksums.

The elf_st_aux structure contains reserved space for other uses not yet
supported by the libelf loader.

Performing normal checksums on the loaded data is slow on most
architectures as it is done by copying the data back out of an ioremapped
range to a kernel buffer and performing the checksum there.  On the ST40
we bypass this as we know ioremapped addresses are the same as kernel
virtual addresses.

Note: checkpatch.pl isn't keen on the __attribute__((packed)) syntax, but I've
kept it for easier merging with the original in the non-GPL OBSP/BSSP sources.

Signed-off-by: David Cook <david.cook@st.com>