stm: mali: pm: fixed pm issue.
The current mali implementation is raising some issue with PM.

The first issue is that the mali_platform_init is called twice;
due to that the clk_prepare_enable is called twice and at the end
the mali clk is enabled forever.

The second issue is that it's not supporting correctly the PM_runtime.
The function mali_platform_power_mode_change isn't managing
the LIGHT_SLEEP power mode (a kind of PM_runtime_suspend).

Due to that a sequence like:

 - LIGHT_SLEEP -> MODE_ON -> LIGHT_SLEEP -> MODE_ON -> ...

increases the mali_clk->usage_counter and again the mali_clk
will never be disabled.

This patch guaratees just one real initialization in the
mali_platform_init and in the mali_platform_power_mode_change
is managing the LIGHT_SLEEP to guarantee balanced operation on
the mali_clk.

Reference: [Bugzilla 22624]

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>