stm: arm: l2x0: add support for stih416 SoC
STiH416 needs some special configuration for TAG and DATA latency as below:

------------------------------------------------------------
                    |  write   |   read   |  setup   | bits
                    | # cycles | # cycles | # cycles |
------------------------------------------------------------
TAG latency cycles  |    3     |    3     |     3    | 0x222
DATA latency cycles |    2     |    2     |     2    | 0x111
-------------------------------------------------------------

Further, as the reset value of the aux ctrl register is 0x200000 (due to
a different set in the RTL of L2$ for Orly-2), it implies that bits[19:17] (way side)
are zero, so it's mandatory to explicitely set the way size by setting bits [19:17]
accordingly. The mask for writing the aux ctrl register must be includes
also the bits for the way size.

(For other chips, reset value is 0x208000, so the way size is correctly set)
Anyway, it's absolutely safe to force this writing for all the chips, without
relying upon the reset value of the control register.

To make code more readable, prefer using meaningful macros, instead of plain
numeric constants.

This patch also fixes the way size for SASC-1, that is 64KB instead of
128KB.

Suggested-by: Aymen Bouattay <aymen.bouattay@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>