stm: b2000: enable IC+ PHY driver
This patch enables the icplus driver by default
otherwise we see the strange phenomenon reported
in the Bugzilla Bug: 25015.
If the GMAC is configure in GMII mode the stmmac
driver always works (100/1000).
If we configure the GMAC in MII mode it only works
if the phy is connected to 10/100 switch. If we
connect the board to a GiGa switch the stmmac fails
when reset the DMA engine.
This happens if we do not enable the ICPLUS driver
so the driver uses the Generic driver provided by the
Physical Layer.
If we configure the ICPLUS all modes work fine.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>