stmmac: start adding pcs and rgmii core irq
This patch starts adding in the main ISR the
management of the PCS and RGMII/SGMII core
interrupts. This is to help further development
on this area.
Currently the core irq handler only clears the
PCS and S-R_MII interrupts and reports the event
in the ethtool stats.

See Bugzilla Bug: 24173

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: Udit Kumar <udit-dlh.kumar@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>