stm: usb_triple_phy: initialized in USB code
This patch changes how the usb_triple_phy is managed.
It was registered as clk but IP-designer confirmed is
a signal (not a clock) which has to be set when the usb_clk is stable.

This patch aligns the usb_triple_phy to what done on Orly-2
Morever it prepares for the common clk framework.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
[stuart.menefy@st.com: Added error detection if cannot claim sysconf]
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>