stm_pcie: Trap aborts to handle UR config requests
On the ARM, if a config request to the PCIE bus returns an UR
(Unsupported Request), then this error is returned to the CPU as a bus
error. With the previous code, this will call the default handler for
this case which will result in a kernel panic. On the ST40, a read with
a bus error returns a value of zero, which is what the existing code
uses to infer that a UR has been returned.

Returing UR is a perfectly legitimate thing to do, it is most commonly
seen when probing a switch that has no device connected on one of it's
ports. Therefore just panicing is not acceptable.

What this patch does is hook the async abort exception handler.
Fortunately it is possible to limit the range of where the exception
will be raised by using the barrier instruction. Thus, when the abort
happens it checks if the PC of the faulting code was where expected.  If
so it sets a flag to indicate the abort has happened which can then be
checked in the code. If it is not in the expected range then the default
panic handler will be called as usual, so any code accessing places it
should not will still be aware of this.

There are two regions delimited, for the config read and write
functions. The regions are marked by inserting asm labels around the
appropriate bits of code. Another assembler macro is used to extract the
address of these labels.

The patch also fixes up an unused variable warning when the code is
compiled for the SH4, due to the ARMs non standard interrupt usage.

Signed-off-by: David McKay <david.mckay@st.com>