stm: hom: updated to be board driven
Several issues were raised from customer on how HoM is managing the
board.

Mainly two issue where raised:
 - customer boards are using a different lmi_retention gpio
 - some board power line has to be enabled or disabled based on which
   devices can wakeup

Due to that this patch is updateing the HoM support to be totally
board driven; from now on the HoM support registration is triggered
from the board file.

Each board (where HoM has to be supported) has to declare
which is the lmi_retention gpio.

Moreover the .freeze/.restore board callback functions now are
aware of which devices can wakeup and based on that the
board-specific code can maintain (or not) some power line enabled.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>