stm_pcie: Rejig label code for abort handling
This patch changes the scheme used in commit
7bd5045639ee732ccd026afd67edb35ee5243dcf to get the address of labels in
the code. The previous scheme used the adr psuedo-op to get the address
of labels in a different function. However, it turns out that adr is too
restrictive in the size of the offset it can encode for this to work
with different kernel configs.

The obvious solution is to replace the adr pseudo-op with the adrl
pseudo-op, as this gives a much larger offset range. Unfortunately this
pseudo op is not available in thumb mode, and I don't want to break the
kernel for thumb mode.

So, what this patch does is rejig the code slightly so that where the
address of the labels is taken is as close as possible to where the
actual label is declared. This ensures that the offset is small and
can easily be encoded by the adr pseudo-op.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>