systrace: add support for the System Trace Module, as part of the MTT infra
The System Trace Module IP exposes basically five or more pages of
4k memory ranges (starting at the base address of the peripheral)
that can be assigned freely to an ARM core, or an accelerator
(XP70 or alike), or peculiar Bus Occupation Measurement IPs.

These pages are typically assigned for use by a core or an accelerator which
is then referred to as an "initiator" or "master" is the STM/STP literature.

Within each page, each 2*64bit aligned offet corresponds to an STP channel.
Writing to BASE + 0x00 will issue a message to Master0/Channel0
Writing to BASE + 0x10 will issue a message to Master0/Channel1

etc..

Odd 8 byte aligned offsets allow to issue a message with an 'XTI' terminator.

Writing to BASE + 0x08 will issue a message to Master0/Channel0 _with_
termination bits.

This can be used by software not dealing with a streaming scheme, that wish
to wait for a specific message termination marker.

The systrace driver in the contrary from other existing implementations
does not provide a directly accessible chardev interface or a commodity
to freely mapp any channel/master to usermode. This is choice of
implementation, as SysTrace is part of the supported media in the MTT
infrastructure. This is to insure:

a) media abstraction in the tools.
b) possible inter-working: trace from one core or app does not break traces
from another.

Documentation: see systrace.txt

Signed-off-by: Marc Titinger <marc.titinger@st.com>