mttS: add a console device to use in place of ttyAS to trace printk to MTT
This allows adding the trace infrastructure to the list of active consoles
used for printk. Passing `console=mttS0` as an additional console to the
kernel make console_unlock/call_console_drivers/ write printk data over
the output medium currently selected for MTT, the default being Systrace.

If ttyAS is still specified in the command line, call_console_drivers will
output the data over it too, so it is recommended removing it. Upon
registration of mttS0, the log_buffer is dumped on this console, so dmesg
level info is made available through the STM-Port aswell in this manner.

Signed-off-by: Marc Titinger <marc.titinger@st.com>