stm_miphy: Allow multi-instance miphy code
This patch changes the register read/write interface to take a struct
stm_miphy *, rather than a port number. The reason for this change is to
allow the read/write functions to be a bit smarter.  For example, later
chips have two uports onto the same phy, one attached to the SATA
controller and one attached to the PCIe controller. The register
read/write functions now have the information available to select the
correct uport to write to depending on which mode the miphy is
configured to be in. The port number is also contained in the miphy
struct, so port selection is again easy.

The other change this patch introduces is to allow the pcie-uport and
the dummy miphy to be multi instance - the existing code could only
handle one instanciation of a phy as it uses static variables. The tap
driver is still only single instance, because the underying tap
interface itself is also only single instance.

This patch also adds a stm_miphy_find_port() function which allows
mapping between a port number and a stm_miphy struct. This is needed for
the peculiar tap case, where in order to initialise a particular port it
is necessary to talk to the other port of the pair.

These changes are to enable support for later chips with more than
one phy present, such as Orly2 and SASC1.

Signed-off-by: David McKay <david.mckay@st.com>